# specs/001-ghostarr-newsletter/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Ghostarr API
  description: Newsletter generator for media servers
  version: 1.0.0
  contact:
    name: Ghostarr
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /newsletters/generate:
    post:
      tags: [newsletters]
      summary: Generate newsletter
      operationId: generateNewsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerationConfig'
      responses:
        '202':
          description: Generation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /newsletters/preview:
    post:
      tags: [newsletters]
      summary: Preview newsletter without publishing
      operationId: previewNewsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerationConfig'
      responses:
        '200':
          description: Preview HTML
          content:
            application/json:
              schema:
                type: object
                properties:
                  html:
                    type: string
                  title:
                    type: string
                  items_count:
                    type: integer
...
  /schedules:
    get:
      tags: [schedules]
      summary: List schedules
      operationId: listSchedules
      responses:
        '200':
          description: Schedule list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScheduleResponse'

    post:
      tags: [schedules]
      summary: Create schedule
      operationId: createSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleResponse'
  /schedules/{id}:
    get:
      tags: [schedules]
      summary: Get schedule
      operationId: getSchedule
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Schedule details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleResponse'

    put:
      tags: [schedules]
      summary: Update schedule
      operationId: updateSchedule
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleResponse'

    delete:
      tags: [schedules]
      summary: Delete schedule
      operationId: deleteSchedule
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
  /schedules/{id}/toggle:
    post:
      tags: [schedules]
      summary: Toggle schedule active status
      operationId: toggleSchedule
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleResponse'
  /schedules/{id}/execute:
    post:
      tags: [schedules]
      summary: Execute schedule immediately
      operationId: executeSchedule
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '202':
          description: Execution started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: array
                items:
                  type: object
                  properties:
                    loc:
                      type: array
                      items:
                        type: string
                    msg:
                      type: string
                    type:
                      type: string
  schemas:
...
    GenerationConfig:
      type: object
      required: [template_id]
      properties:
        template_id:
          type: string
          format: uuid
        title:
          type: string
          default: "Newsletter {{date.week}}"
        publication_mode:
          type: string
          enum: [draft, publish, email, email+publish]
          default: draft
        ghost_newsletter_id:
          type: string
          nullable: true
        tautulli:
          $ref: '#/components/schemas/TautulliConfig'
        romm:
          $ref: '#/components/schemas/ContentSourceConfig'
        komga:
          $ref: '#/components/schemas/ContentSourceConfig'
        audiobookshelf:
          $ref: '#/components/schemas/ContentSourceConfig'
        tunarr:
          $ref: '#/components/schemas/TunarrConfig'
        statistics:
          $ref: '#/components/schemas/StatisticsConfig'
        maintenance:
          $ref: '#/components/schemas/MaintenanceConfig'
        max_total_items:
          type: integer
          default: -1
        skip_if_empty:
          type: boolean
          default: false
    ContentSourceConfig:
      type: object
      properties:
        enabled:
          type: boolean
          default: false
        days:
          type: integer
          minimum: 1
          maximum: 90
          default: 7
        max_items:
          type: integer
          minimum: -1
          default: -1
    TautulliConfig:
      allOf:
        - $ref: '#/components/schemas/ContentSourceConfig'
        - type: object
          properties:
            featured_item:
              type: boolean
              default: false
    TunarrConfig:
      allOf:
        - $ref: '#/components/schemas/ContentSourceConfig'
        - type: object
          properties:
            days:
              type: integer
              minimum: 1
              maximum: 7
              default: 7
            channels:
              type: array
              items:
                type: string
            display_format:
              type: string
              enum: [grid, list]
              default: grid
    StatisticsConfig:
      type: object
      properties:
        enabled:
          type: boolean
          default: false
        days:
          type: integer
          minimum: 1
          maximum: 90
          default: 7
        include_comparison:
          type: boolean
          default: false
    MaintenanceConfig:
      type: object
      properties:
        enabled:
          type: boolean
          default: false
        description:
          type: string
        type:
          type: string
          enum: [scheduled, outage, network, update, improvement, security]
          default: scheduled
        duration_value:
          type: integer
          default: 1
        duration_unit:
          type: string
          enum: [hours, days, weeks]
          default: hours
        start_datetime:
          type: string
          format: date-time
          nullable: true

    # ---- Templates ----
    TemplateResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        file_path:
          type: string
        preset_config:
          $ref: '#/components/schemas/GenerationConfig'
        is_default:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ScheduleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        is_active:
          type: boolean
        cron_expression:
          type: string
        cron_human_readable:
          type: string
        timezone:
          type: string
        template_id:
          type: string
          format: uuid
        template:
          $ref: '#/components/schemas/TemplateResponse'
        generation_config:
          $ref: '#/components/schemas/GenerationConfig'
        last_run_at:
          type: string
          format: date-time
          nullable: true
        last_run_status:
          type: string
          enum: [pending, success, failed, skipped]
          nullable: true
        next_run_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ScheduleCreate:
      type: object
      required: [name, cron_expression, template_id, generation_config]
      properties:
        name:
          type: string
        cron_expression:
          type: string
        timezone:
          type: string
          default: Europe/Paris
        template_id:
          type: string
          format: uuid
        generation_config:
          $ref: '#/components/schemas/GenerationConfig'
    ScheduleUpdate:
      type: object
      properties:
        name:
          type: string
        cron_expression:
          type: string
        timezone:
          type: string
        template_id:
          type: string
          format: uuid
        generation_config:
          $ref: '#/components/schemas/GenerationConfig'
        is_active:
          type: boolean

    # ---- History ----
    HistoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [manual, automatic]
        schedule_id:
          type: string
          format: uuid
          nullable: true
        schedule:
          $ref: '#/components/schemas/ScheduleResponse'
        template_id:
          type: string
          format: uuid
        template:
          $ref: '#/components/schemas/TemplateResponse'
        status:
          type: string
          enum: [pending, running, success, failed, cancelled]
        ghost_post_id:
          type: string
          nullable: true
        ghost_post_url:
          type: string
          nullable: true
        generation_config:
          $ref: '#/components/schemas/GenerationConfig'
        progress_log:
          type: array
          items:
            $ref: '#/components/schemas/ProgressStep'
        error_message:
          type: string
          nullable: true
        items_count:
          type: integer
        duration_seconds:
          type: number
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...
    ProgressStep:
      type: object
      properties:
        step:
          type: string
        status:
          type: string
          enum: [pending, running, success, failed, skipped]
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        duration_ms:
          type: integer
        items_count:
          type: integer
        message:
          type: string
        error:
          type: string

    # ---- Settings ----
...